rename similarities.index to similarities.annoy
